Refactor to reduce duplication
Move process queue to Abscract Process class
